Ensure request payload IO object is read for reading
Internally rest-client stores the payload as an IO object, so when
api_auth reads it to calculate an MD5, it moves the pointer position to
the end of the IO object. Calling inspect resets the pointer position
back to 0 so it can be read again when sending the actual HTTP request.